feat!: support all major debian/ubuntu versions
There's a lot going on here...

Context
=======
This commit represents a resumed effort;
I first started this project nearly a year ago (April/May 2022).

When I left off, I was mid-refactor/reorganization, so this is my
best-effort at recapturing the spirit.

Decisions
=========
I went back and forth on whether or not to have a combined index for
Debian and Ubuntu, with each selecting the appropriate codename
subdirectory.

My thinking was that there may be some esoteric edge cases worth
implicitly allowing, namely mixing and matching dist components between
Debian and Ubuntu.

It would also make bootstrapping a little easier,
as we'd use a shared URL for the `Bootstrap` script.

Unfortunately, I've come to decide it's not worth the hassle.

Drawbacks
---------
Generally, combining the two makes things messier.

- The `dists` directory contains twice as many entries,
  with no indication to which distro they belong.
- Much existing tooling expects the two fields `{ID}-{CODENAME}` to be
  present, so any time we diverge from this, we need to write more
  support code.

While presumably _highly_ unlikely, it's possible that a future
distro could use a codename already used by the other. The odds of this
increase, the more Debian variants we include in this repository.

There also exist enough differences between the two package managers
that we benefit from a cleaner separation at the onset.

- During development of this patch, I learned that the default packages
  built _with_ `equivs-build` on _Ubuntu_ were creating
  Debian-incompatible packages.
    - Ubuntu's version compresses the package contents with `zstd`,
      but this isn't added to Debian's `apt` until v12.
        - We now recompile all metapackage .debs to ensure they are
          compressed with `gzip`, for the broadest compatibility.
          It's also what we were explicitly using everywhere else.

To avoid future issues, we're keeping the two separate for now.

Makefiles
=========
These are highly dynamic and have been refactored to be generalized.
A common `Makefile` is shared between siblings, ie. debian+ubuntu,
jammy+bionic+focal+buster+bullseye+stretch.